Wire Model Analysis as 6th drawer tab
Splits app/page.tsx into a thin Server Component that calls
readManifest() and a renamed app/HomeClient.tsx that holds the
existing client UI (now accepting manifestRuns as a prop and
rendering a 6th "Model Analysis" tab).

Also updates app/about/page.tsx, which used the old default
Home export as a decorative background, to import HomeClient
and pass manifestRuns={[]} (the empty state is harmless for
an aria-hidden background).
